Test repo showing instrumenter issue
Run with `npm run test` to produce two files, "statementMap_nyc.txt" and "statementMap_istanbul.txt".
The test command will compare the instrumenter output from nyc@latest and istanbul@0.4.5.
The output here from nyc has an extra statement that corresponds to a character of whitespace in the initialisation of a for loop.